POS-264 - Introduce a directive to more easily add a 'BETA' flag to document title (and add it to the Terminals API)
This took me way more time than I'd like to admit, but at least now we no longer have to fiddle with raw HTML any longer.

Let's say the Sphinx / ReStructuredText documentation is not that extensive to take into consideration use-cases like these.

The directive looks for the title node, and adds the 'BETA' flag to it. I added this directive to the overview pages of the Balances and Terminals API.