POS-266 - Move POS integration guide to its own directory, as we'll be adding a page on testmode
Currently, the POS integration guide was part of the 'payments' source directory. Since 'Point of Sale' has its own section in the sidebar, and we're now going to be adding a second Point of Sale related page, it made more sense to me to group these pages in a separate directory.